Commit #29
The error is showing “Namespace 'app' is not bound” indicates that the app namespace is not declared at the top of your XML layout file. To fix this, you can add the xmlns:app declaration at the start of the file. Here’s the updated XML with the missing app namespace added